pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssg"
version = "0.1.0"
description = "A simple, lightweight static site generator built with Python"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
dependencies = [
    "jinja2>=3.0.0",
    "markdown>=3.3.0",
    "python-frontmatter>=1.0.0",
]

[project.scripts]
ssg = "ssg.cli:main" 